apps/serializers/product.py
from rest_framework.serializers import ModelSerializer

from apps.models import Product, Category


class ProductModelSerializer(ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'


class CategoryModelSerializer(ModelSerializer):
    class Meta:
        model = Category
        fields = '__all__'
