pyproject.toml
[project]
name = "sof-mebel"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "django-filter>=25.2",
    "django-jazzmin>=3.0.4",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.29.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
]
